change renewLock signature (WP-1001)

+++ inc/Smartling/ApiWrapperInterface.php
@@ -55,7 +55,7 @@ interface ApiWrapperInterface
     /**
      * @throws SmartlingApiException
      */
-    public function renewLock(ConfigurationProfileEntity $profile, string $key, int $ttlSeconds): \DateTime;
+    public function renewLock(ConfigurationProfileEntity $profile, string $key, float $ttlSeconds): \DateTime;
 
     /**
      * @throws SmartlingApiException
