+++ inc/Smartling/ApiWrapperWithRetries.php
@@ -37,7 +37,7 @@ class ApiWrapperWithRetries implements ApiWrapperInterface {
         });
     }
 
-    public function renewLock(ConfigurationProfileEntity $profile, string $key, int $ttlSeconds): \DateTime
+    public function renewLock(ConfigurationProfileEntity $profile, string $key, float $ttlSeconds): \DateTime
     {
         return $this->withRetry(function () use ($profile, $key, $ttlSeconds) {
             return $this->base->renewLock($profile, $key, $ttlSeconds);
